feat: add bar chart
